docker/fastapi/requirements.txt
fastapi[all]
uvicorn[standard]
python_dotenv
openai